% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{per.cover.data}
\alias{per.cover.data}
\title{Percent cover data for benthic categories}
\format{
A data frame where each row corresponds to a quadrat image and
columns represent percent cover of benthic categories.
}
\source{
CoralNet annotations of coral reef images
}
\usage{
data(per.cover.data)
}
\description{
This dataset contains percent cover estimates of benthic functional groups
and taxa derived from CoralNet image analysis.
}
\keyword{datasets}
